+++ infra/orbstack-local/providers.tf
@@ -4,3 +4,8 @@ provider "helm" {
     config_context = "orbstack"
   }
 }
+
+provider "kubernetes" {
+  config_path    = "~/.kube/config"
+  config_context = "orbstack"
+}